pyproject.toml
[project]
name = "dio-blog"
version = "0.1.0"
description = ""
authors = [
    {name = "Paulo Santos"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.116.0,<0.117.0)",
    "databases[aiosqlite] (>=0.9.0,<0.10.0)",
    "black (>=25.1.0,<26.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
]



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
httpx = "^0.28.1"
pytest-mock = "^3.14.1"


[tool.pytest.ini_options]
asyncio_mode = "auto"
